pyproject.toml
[tool.poetry]
name = "vuanem_basevn"
version = "2.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
google-cloud-bigquery = "^2.30.1"
google-cloud-tasks = "2.5.1"
httpx = "^0.22.0"
compose = "^1.2.8"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
pytest = "^6.2.5"
pytest-timeout = "^2.0.1"
mypy = "^0.910"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"
timeout = 530
testpaths = [
    "test",
]
